# SwaggerDoc.yml
openapi: 3.0.0
info:
  version: "0.3.0"
  title: signer
...
paths:
...
  "/tokens/:id":
    get:
      description: Validate and decode a token
      tags:
        - /tokens
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Token valid
          content:
            application/json:
              schema:
                type: object
                example: { "name": "Simon", ...: "Your Payload Data" }
                
        401:
          description: Token invalid
          content:
            text/plain:
              schema:
                type: string
    
    head:
      tags:
        - /tokens
      description: Validate a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Token valid
        401:
          description: Token invalid
    
    delete:
      tags:
        - /tokens
      description: Block a token
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Token Blocked